extract global ui
